requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email-validator]==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
google-generativeai==0.8.3
openpyxl==3.1.5
python-multipart==0.0.12
